pyproject.toml
# SPDX-License-Identifier: MIT

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "changwon-corp-credit"
version = "0.1.0"
description = "Changwon corporate credit analysis pipeline for Doosan Enerbility."
authors = [
  { name = "Codex", email = "codex@example.com" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "dash>=2.18.0",
  "dash-bootstrap-components>=2.0.0",
  "lxml>=5.2.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "pyarrow>=16.0.0",
  "pyyaml>=6.0.0",
  "requests>=2.32.0",
  "rich>=13.7.0",
  "typer>=0.12.0",
  "plotly>=5.24.0",
  "kaleido==0.2.1",
  "streamlit>=1.39.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-mock>=3.12.0",
]

[project.scripts]
changwon-credit = "changwon_credit.cli:app"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/changwon_credit"]

[tool.hatch.build.targets.sdist]
include = [
  "src/changwon_credit",
  "README.md",
  "config",
  "schema",
  "scripts",
  "tests",
]
